% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upre.R
\name{interact}
\alias{interact}
\title{Calculate interaction test statistic for user-specified variable}
\usage{
interact(object, varnames, nullmods = NULL, k = 10, plot = TRUE,
  col = c("red", "yellow"), ylab = "Interaction strength", ...)
}
\arguments{
\item{object}{an object of class \code{\link{upre}}.}

\item{varnames}{character vector. Names of variables for which interation test 
statistic should be calculated.}

\item{nullmods}{object with bootstrapped null interaction models, resulting 
from application of \code{bsnullinteract}.}

\item{k}{integer. Calculating interaction test statistics is a computationally 
intensive, so  calculations are split up in several parts to prevent memory 
allocation errors. If a memory allocation error still occurs, increase k.}

\item{plot}{logical Should the interaction statistics be plotted?}

\item{col}{character vector of length two. Color for plotting bars used. Only 
used when plot = TRUE. Only first element of vector is used if 
\code{nullmods = NULL}.}

\item{ylab}{character string. Label to be used for plotting y-axis.}

\item{...}{Additional arguments to be passed to \code{\link[graphics]{barplot}}}
}
\value{
If nullmods is not specified, the function returns the test statistic 
of the interaction. If nullmods is specified, the function returns a list, 
with elements \code{$H}, which is the test statistic of the interaction 
strength, and \code{$nullH}, which is a vector of test statistics of the 
interaction in each of the bootstrapped null interaction models.
}
\description{
\code{interact} calculates a statistic for assessing the presence of 
interactions between the input variable(s) specified, and all other input
variables.
}
\details{
Can be computationally intensive, especially when nullmods is specified.
}
\examples{
\dontrun{
 airq.ens <- upre(Ozone ~ ., data=airquality[complete.cases(airquality),])
 interaction(airq.ens, "Temp")}
}

